+++ app/Http/Controllers/FamilyController.php
@@ -17,7 +17,11 @@ class FamilyController extends Controller
     {
         //
         $families = Family::all();
-        return view('families.index', compact('families'));
+        foreach ($families as $index=>$family):
+          $families[$index]['socialLinks'] = $family->socialLinks;
+          $families[$index]['performers'] = $family->performers;
+        endforeach;
+        return response()->json($families, 200);
     }
 
     /**
@@ -46,8 +50,14 @@ class FamilyController extends Controller
           'description' => 'required',
         ]);
         $family = Family::create($attributes);
-        $family_id = $family->id;
-        return view('socialLinks.create', compact('family_id'));
+        $user = Performer::find(request('user')->performer['id']);
+        $family->performers()->save($user);
+        $performers = request('performers');
+        foreach ($performers as $performer):
+          $newPerformer = Performer::find($performer);
+          $family->performers()->save($newPerformer);
+        endforeach;
+        return response()->json(['status' => 'success'], 200);
     }
 
     /**
@@ -89,11 +99,17 @@ class FamilyController extends Controller
      * @param  \App\Family  $family
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Family $family)
+    public function update($id)
     {
         //
-        $family->update(request(['name', 'description']));
-        return redirect('/families/'.$family->id);
+        $family = Family::find($id);
+        $user = request('user');
+        $userPerformer = $user->performer;
+        if ($userPerformer['family_id'] === $family['id']):
+          $family->update(request(['name', 'description']));
+          return response()->json(['status' => 'success'], 200);
+        endif;
+        return response()->json(['message' => 'unauthorized'], 405);
     }
 
     /**
@@ -102,34 +118,47 @@ class FamilyController extends Controller
      * @param  \App\Family  $family
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Family $family)
+    public function destroy($id)
     {
         //
-        $performers = $family->performers;
-        foreach($performers as $performer) {
-          $performer->family_id = null;
-          $performer->save();
-        }
-        $family->delete();
-        return redirect('/families');
+        $family = Family::find($id);
+        $user = request('user');
+        $userPerformer = $user->performer;
+        if ($userPerformer['family_id'] === $family['id']):
+          $performers = $family->performers;
+          foreach($performers as $performer) {
+            $performer->family_id = null;
+            $performer->save();
+          }
+          $family->delete();
+          return response()->json(['status' => 'success'], 200);
+        endif;
+        return response()->json(['status' => 'unauthorized'], 401);
     }
 
     public function performer($id) {
       $family = Family::find($id);
       $performer = Performer::find(request('performer'));
-      $family->performers()->save($performer);
-      $allPerformers = Performer::all();
-      $familyPerformers = Family::find($id)->performers;
-      return redirect('/families/'.$family->id.'/edit')->with(compact('allPerformers', 'familyPerformers', 'family'));
+      $userPerformer = request('user')->performer;
+      if (intval($id) === intval($userPerformer['family_id'])):
+        $family->performers()->save($performer);
+        $allPerformers = Performer::all();
+        $familyPerformers = Family::find($id)->performers;
+        return response()->json(['status' => 'success'], 200);
+      endif;
+      return response()->json(['message' => 'unauthorized user'], 401);
     }
+
     public function performerDestroy($id) {
       $performer = Performer::find($id);
-      $family_id = $performer->family_id;
-      $performer->family_id = null;
-      $performer->save();
-      $allPerformers = Performer::all();
-      $family = Family::find($family_id);
-      $familyPerformers = $family->performers;
-      return redirect('families/'.$family_id.'/edit')->with(compact('allPerformers', 'familyPerformers', 'family'));
+      $family = $performer->family;
+      $userPerformer = request('user')->performer;
+      if (intval($family['id']) === intval($userPerformer['family_id'])):
+        $performer->family()->dissociate();
+        $performer->save();
+        return response()->json(['status' => 'success'], 200);
+      endif;
+      return response()->json(['message' => 'unauthorized user'], 401);
+
     }
 }
